+++ app/client/nfts/[id]/page.tsx
