+++ app/client/shows/[id]/page.tsx
