+++ app/client/shows/page.tsx
