+++ app/globals.css
@@ -2,11 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 /* sm: >= 640px，md: >= 768px，lg: >= 1024px，xl: >= 1280px */
-
-body {
-  background: #020202;
-  box-sizing: border-box;
-}
 @layer base {
   :root {
     --background: 0 0% 100%;
