+++ app/page.tsx
@@ -1,16 +1,4 @@
-import MaxWidthWrapper from "@/components/MaxWidthWrapper";
-import Hero from "@/components/client/hero/Hero";
-import Header from "@/components/header/Header";
-
+import { redirect } from "next/navigation";
 export default function Home() {
-  return (
-    <>
-      <Hero />
-      <MaxWidthWrapper className="mx-auto text-white">
-        <Hero />
-        <Hero />
-        <Hero />
-      </MaxWidthWrapper>
-    </>
-  );
+  return redirect("/client");
 }
