+++ components/client/MaxWidthWrapper.tsx
@@ -8,7 +8,7 @@ const MaxWidthWrapper = ({ children, className }: IMaxWidthWrapper) => {
   return (
     <section
       className={cn(
-        "mx-auto w-full max-w-screen-2xl px-5 lg:px-10 ",
+        "mx-auto w-full max-w-screen-2xl px-5 md:px-10 lg:px-20 ",
         className
       )}
     >
