+++ components/client/header/Header.tsx
@@ -4,12 +4,14 @@ import { Menu, Webhook } from "lucide-react";
 import NavItem from "./NavItem";
 import { CustomConnectButton } from "./CustomConnectButton";
 import MaxWidthWrapper from "../MaxWidthWrapper";
-import { ConnectButton } from "@rainbow-me/rainbowkit";
+
+import { cn } from "@/lib/utils";
+import { paths } from "@/utils/paths";
 const Header = memo(function Header() {
   const navList = [
     {
       title: "Events",
-      link: "/events",
+      link: paths.client.market,
     },
     {
       title: "Collectables",
@@ -29,8 +31,13 @@ const Header = memo(function Header() {
     },
   ];
   return (
-    <header className="backdrop-blur-md w-full fixed top-0 inset-x-0 z-50  py-4 px-6 lg:px-18 mb-3 lg:h-auto">
-      <MaxWidthWrapper className="mx-auto relative flex-1 px-0">
+    <header
+      className={cn(
+        "w-full sticky top-0 inset-x-0 z-50  py-4 px-6 lg:px-18 mb-3 lg:h-auto duration-300 backdrop-blur-md"
+      )}
+      id="scroll-bar"
+    >
+      <MaxWidthWrapper className="flex-1 relative px-0 md:px-0 lg:px-0">
         <div className="flex justify-between items-center ">
           <Link
             href="/"
@@ -41,7 +48,7 @@ const Header = memo(function Header() {
           </Link>
           {/* Mobile */}
           <div className="hover:cursor-pointer h-full flex lg:hidden text-orange-500">
-            <Menu size={40} />
+            <Menu size={40} color="#fff" />
           </div>
         </div>
         <div className="hidden lg:flex flex-col justify-end right-0 lg:mt-0 lg:absolute lg:top-0 lg:h-full lg:items-center">
