+++ package.json
@@ -13,6 +13,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@rainbow-me/rainbowkit": "^2.0.6",
     "@tanstack/react-query": "^5.32.0",
+    "axios": "^1.6.8",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "ethers": "^6.12.1",
