+++ utils/paths.ts
@@ -1,6 +1,9 @@
 // api v1
 export const paths = {
-  home: "/",
-  market: "/market",
-  create: "/create",
+  client: {
+    home: "/client",
+    market: "/market",
+    create: "/create",
+  },
+  admin: {},
 };
